Start of changeover to pattern-based navigation
Ingesting the new JSON files for pattern info and star system info. Not yet using star system info data for anything. For the pattern data, can press spacebar to switch to next pattern and see the pattern name and the name of the first star system that has it. The planets rendered just stay on solar system planets for now.