Can display solar system using Tab key
In-Progress #1 Need to add being able to go back to the pattern you were on when you jumped. For now, you can navigate to the pattern before or after the one you were on. Also note that the display of the solar system is such that the larger planets are covering up some text elements, so will need to adjust canvas size/content spacing.