HUGE overhaul to make planets 3D via WEBGL
Now using p5js in instance mode, which means there are multiple p5 sketches in the app. This allows me to have a sketch using a 3D canvas and put a 2D canvas on top. The 3D canvas is only rendering the planets (as spheres) and also has ambient and directional lighting. Colored the directional lighting to match the color of the star (eventually will have the star change color per system). HTML and CSS changes were required to enable the canvases to be directly on top of each other. Fixes #24